vuelo variables de transicion

+++ kernel/src/config.c
@@ -15,7 +15,7 @@ t_kernel_config* ini_kernel_config(char* path_a_config) {
     kernel_config -> estimacion_inicial = config_get_int_value(kernel_config -> config, "ESTIMACION_INICIAL");
     kernel_config -> alfa = config_get_int_value(kernel_config -> config, "ALFA");
     kernel_config -> grado_multiprogramacion = config_get_int_value(kernel_config -> config, "GRADO_MULTIPROGRAMACION");
-    kernel_config -> tiempo_maximo_bloqueado = (double) config_get_int_value(kernel_config -> config, "TIEMPO_MAXIMO_BLOQUEADO");
+    kernel_config -> tiempo_maximo_bloqueado = config_get_int_value(kernel_config -> config, "TIEMPO_MAXIMO_BLOQUEADO");
 
     return kernel_config;
 }
